requirements.txt
Django>=1.4
Markdown>=2.1.1
South>=0.7.5
akismet>=0.2.0
argparse>=1.2.1
django-bootstrap-pagination>=0.1.10
django-disqus>=0.4.1
django-pagination>=1.0.7
django-tagging>=0.3.1
django-tinymce>=1.5.1b2
wsgiref>=0.1.2
